services/audio-production/requirements.txt
# Core dependencies
torch>=2.2.0
torchaudio>=2.2.0
numpy>=1.24.0
scipy>=1.11.0

# Audio processing
librosa>=0.10.1
soundfile>=0.12.1
pydub>=0.25.1
pyloudnorm>=0.1.1

# Audio generation (install without dependencies to avoid version conflicts)
audiocraft>=1.3.0
requests  # For Suno API HTTP calls
demucs>=4.0.0

# API framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
httpx>=0.26.0
python-dotenv>=1.0.0
supabase>=2.3.0
